pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastmvc-middleware"
version = "0.2.0"
description = "Production-ready middleware collection for FastAPI/Starlette applications"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Shiv", email = "shiv@hyyre.dev" }
]
keywords = [
    "fastapi",
    "starlette",
    "middleware",
    "cors",
    "security",
    "rate-limiting",
    "authentication",
    "logging",
    "api",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "starlette>=0.27.0",
]

[project.optional-dependencies]
jwt = ["pyjwt>=2.0.0"]
all = [
    "pyjwt>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pyjwt>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]

[project.urls]
Homepage = "https://github.com/hyyre/fastmvc-middleware"
Documentation = "https://github.com/hyyre/fastmvc-middleware#readme"
Repository = "https://github.com/hyyre/fastmvc-middleware"
Issues = "https://github.com/hyyre/fastmvc-middleware/issues"
Changelog = "https://github.com/hyyre/fastmvc-middleware/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.sdist]
include = [
    "/fastMiddleware",
]

[tool.hatch.build.targets.wheel]
packages = ["fastMiddleware"]
sources = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --cov=fastMiddleware --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["fastMiddleware"]

[tool.coverage.run]
source = ["fastMiddleware"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

